selectTaskForCurrentExploreSection
