+++ src/selectors/tasks/select_task_for_current_explore_section.ts
@@ -1,0 +1,28 @@
+import * as R from 'ramda';
+import { Store } from '../../stores';
+import * as store from '../../stores/tasks';
+import { selectTaxonomyTermsForChosenAnswers } from '../taxonomies/select_taxonomy_terms_for_chosen_answers';
+import { RouterProps } from '../../application/routing';
+import { selectLocale } from '../locale/select_locale';
+import { findItemByLearnTaxonomyTerm } from '../taxonomies/find_item_by_explore_taxonomy_term';
+import { Task } from './task';
+import { toSelectorTaskWithoutRelatedEntities } from './to_selector_task_without_related_entities';
+import { selectExploreSectionFromTask } from './select_explore_section_from_task';
+import { isTaskRecommended } from '.';
+
+export const selectTaskForCurrentExploreSection = (appStore: Store, routerProps: RouterProps): ReadonlyArray<Task> => {
+    const exploreSection = appStore.exploreSectionsInStore.sections[routerProps.match.params.learnId];
+    const tasks = appStore.tasksInStore.taskMap;
+    const matchingTasks = findItemByLearnTaxonomyTerm(exploreSection.taxonomyTerms, tasks);
+
+    const locale = selectLocale(appStore);
+
+    const buildTask = (task: store.Task): Task => {
+        const exploreSectionForTask = selectExploreSectionFromTask(appStore, task);
+        const termsFromQuestionnaire = selectTaxonomyTermsForChosenAnswers(appStore);
+        const isRecommended = isTaskRecommended(termsFromQuestionnaire, task);
+        return toSelectorTaskWithoutRelatedEntities(locale, task, exploreSectionForTask, isRecommended);
+    };
+
+    return R.map(buildTask, matchingTasks);
+};
